+++ src/bindings/matlab/StructureFields.h
@@ -1,84 +1,7 @@
-#ifndef TYPES_CODE_INCLUDED
-#define TYPES_CODE_INCLUDED
+#ifndef SF_INCLUDED
+#define SF_INCLUDED
 
-class ModelDetails;
-struct GV
-{
-	mxArray* modelArray = NULL;
-	bool freeMemory;
-	ModelDetails* details;
-
-	IdList reqdPkgPrefixes;
-	IdList unreqdPkgPrefixes;
-
-	bool fbcUsingId;
-	bool fbcAddGeneProducts;
-	bool onlyExpectedFields;
-	bool applyUserValidation;
-};
-
-
-typedef std::map<const std::string, unsigned int> PkgMap;
-typedef PkgMap::iterator  PkgIter;
-
-
-typedef enum
-{
-    TYPE_BOOL
-  , TYPE_CHAR
-  , TYPE_DOUBLE
-  , TYPE_INT
-  , TYPE_ELEMENT
-  , TYPE_UINT
-  , TYPE_UNKNOWN
-} FieldType_t;
-
-
-const char* FIELDTYPE_STRINGS[] =
-{
-      "bool"
-    , "char"
-    , "double"
-    , "int"
-    , "structure"
-    , "uint"
-};
-
-
-typedef enum {
-    OTHER_NAME = 0
-  , SBML_NOTES
-  , SBML_ANNOT
-  , OTHER_NAMESPACES
-  , OTHER_CVTERMS
-  , SBML_MATH
-  , SBML_RULE_TYPE
-  , SBML_ISSET
-  , SBML_MESSAGE
-  , OTHER_TIME_SYMBOL
-  , OTHER_DELAY_SYMBOL
-  , OTHER_AVOGADRO_SYMBOL
-  , OTHER_RATEOF_SYMBOL
-  , OTHER_ANOMALOUS_MATH
-  , FBC_ASSOCIATION
-
-  , SBML_KNOWN_ATT_ELEM = 20
-
-} FieldnamesType_t;
-
-
-struct FieldValues_t {
-  std::string fieldName;
-  std::string sbmlName;
-  std::string prefix;
-  FieldType_t type;
-  bool isSBMLAttribute;
-  FieldnamesType_t fieldnameType;
-  std::string strValue;
-  int iValue;
-  double dValue;
- };
-
+#include "Variables.h"
 
 class StructureFields
 {
@@ -231,63 +154,4 @@ protected:
   std::vector<FieldValues_t> mFields;
 };
 
-
-class ModelDetails
-{
-public:
-  ModelDetails(GV &gv);
-
-  ModelDetails(SBMLDocument *doc, GV& gv);
-
-  ~ModelDetails() { };
-
-  unsigned int getLevel() {return mLevel; } ;
-  unsigned int getVersion() { return mVersion; } ;
-
-  const std::string& getDelaySymbol() { return mDelaySymbol; } ;
-  const std::string& getTimeSymbol() { return mTimeSymbol; } ;
-  const std::string& getAvogadroSymbol() { return mAvogadroSymbol; } ;
-  const std::string& getRateOfSymbol() { return mRateOfSymbol; } ;
-
-  void setDelaySymbol(const std::string& symbol) { mDelaySymbol = symbol; } ;
-  void setAvogadroSymbol(const std::string& symbol) { mAvogadroSymbol = symbol; } ;
-  void setTimeSymbol(const std::string& symbol) { mTimeSymbol = symbol; } ;
-  void setRateOfSymbol(const std::string& symbol) { mRateOfSymbol = symbol; } ;
-
-  SBMLNamespaces* getNamespaces() { return mSBMLns; };
-
-  PkgMap getPackages() { return mPackageMap; };
-
-  bool isPackagePresent(const std::string& pkg);
-
-
-protected:
-
-  void populateNamespaces();
-  void populateNamespaces(SBMLDocument* doc);
-  void populatePkgMap();
-  void populatePkgMap(SBMLDocument* doc);
-  void populateSupportedPackages();
-
-  bool isSupportedPackageNS(const std::string& uri, const std::string prefix);
-
-  unsigned int mLevel;
-  unsigned int mVersion;
-
-  std::string mDelaySymbol;
-  std::string mTimeSymbol;
-  std::string mAvogadroSymbol;
-  std::string mRateOfSymbol;
-
-  SBMLNamespaces *mSBMLns;
-
-  IdList mSupportedPackages;
-
-  PkgMap mPackageMap;
-
-  GV& gv;
-
-};
-
-
-#endif // TYPES_CODE_INCLUDED
+#endif // SF_INCLUDED
